send mails for payment initiation and payment success and failure

+++ utility-billing-system/payment-service/src/main/java/com/utilitybilling/paymentservice/model/Invoice.java
@@ -4,21 +4,22 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Data
-@Document(collection="invoices")
-public class Invoice{
+@Document(collection = "invoices")
+public class Invoice {
 
-    @Id
-    private String id;
+	@Id
+	private String id;
 
-    private String paymentId;
-    private String billId;
-    private String consumerId;
+	private String paymentId;
+	private String billId;
+	private String consumerId;
 
-    private double amount;
-    private PaymentMode mode;
+	private BigDecimal amount;
+	private PaymentMode mode;
 
-    private Instant generatedAt=Instant.now();
+	private Instant generatedAt = Instant.now();
 }
